Protection

+++ app/Http/Controllers/Admin/PlanController.php
@@ -76,12 +76,20 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+        ->with('details')
+        ->where('url', $url)
+        ->first();
 
         if(!$plan)
             return redirect()->back();
-        $plan->delete();
 
+        if($plan->details->count() > 0){
+            return redirect()->back()->with('error', 'Existem detalhes vinculados a essa plano');
+        }
+        
+        $plan->delete();
+ 
         return redirect()->route('plans.index');
     }
